# perf(utils): use `| 0` instead of `Math.floor`